webapp/requirements.txt
streamlit==1.*
pandas==1.*
plotly==5.*
google-cloud-bigquery==3.* 